Promote e2e-tasmax into e2e-tasmax-jobs template, refactor
"e2e-tasmax" will see lots of play, so we're promoting it to a full WorkflowTemplate. This makes it easier to reuse and keep track of its resource use.

There are some moderate parameters getting refactored in this commit. This is so that the things that make a "tasmax" run a "tasmax" run are set as parameters in the e2e workflowtemplate. Previously, we just relied on default parameters to set tasmax-appropriate values. The change here make parameters more explicit. This will make thing cleaner when we start running other parameters.