Fix precip workflows killed before finishing
Extends workflow deadline from 8 hours (default) to 12 hours. The addition of "ceiling threshold" steps to biascorrection and downscaling parts of the precipitation workflow are new and still inefficient. This extends the workflows deadline so that workflows will not need to be manually retried or restarted.

Ideally, we can polish up these new ceiling steps so they run faster.