Add workflowtemplate to create precip qplad/qdm ref data
This workflowtemplate takes cleaned ERA-5 precipitation data as inputs and outputs the reference for QDM biascorrection, and the fine and coarse references for QPLAD downscaling.

This workflow is intended to be run independently of the end-to-end or biascorrect-downscaling workflows, before those workflows are run. This makes those common runs more efficient. Hopefully, this also makes the logic to create the reference data more clear.

Note that this WorkflowTemplate applies precipitation wet-day frequency correction to the data *after* regridding, as opposed to before regridding. Also, the coarse and QDM reference data share the same wet-day corrections, so that they are the same randomly generated numbers in the correction. This is different from how wet-day correction and references have been calculated prior.