Dynamically detect non-time dims in regrid
Fixes failing distributed regrid regional writes to zarr store when input data has variable that is not "lat" or "lon" and does not depend on "time" dim. The error appears because all data regionally written to the zarr needs to depend on "time". The solution is to dynamically figure out which variables do or do not depend on time. Previously, this depended on a static, user-provided list of variables to exclude.

close #494