Fix bad city position from lon360
Updated notebook and reran daily and clean-daily diagnostics to fix
error. Cities timeseries in diagnostics were not correctly placed when
diagnostics were run with climate fields using longitude from [0, 360].
This also adds a tolerance so an error is thrown if the search
distance is large.

Also includes updates so this thing could run on the
pangeo-notebook:2023.03.27 images.